Engine: generic way for loading save with less content
This allows engine to load saves that were written by (presumably older) game version which had less content (number of entities or size of global script data).
Since not every object can be securely identified at this point, only the sizes of entity arrays are compared.

This is indeed a partial solution, but it may work for starters, before we implement reliable identifiers for all object types (we may do that in steps too).