Engine: implemented "resolve_restored_game" script hook
Resolve_restored_game is a script function that is called whenever a save was loaded, but before the process gets finalized. This means that restoration still can be cancelled at this point.

This script function has only one parameter, which is a LoadedSaveInfo object, which purpose is to describe the contents of the loaded save, help detect and resolve any content conflicts (i.e. when save is made in different version of the game).