Engine: starting to implement new save format
The new format is supposed to consist of self-describing, unordered blocks of data, instead of single stream of predefined data.
The main reasons for doing so, besides cleaning the old savegame routine, is to ease future data changes, simplify savegame debugging and support for auxiliary or optional data in the save.

This commit, in particular, defines new save version, specifies save block structure and writes save description as one obligatory block.