Engine: recreated ScriptUserObject into CCDynamicStruct manager
This makes the former "user object" class not store any data in itself, and instead store additional fields as a header prepended to the struct's raw data. Now DynamicStruct and DynamicArray act practically the same, with the only difference of array having different meta-data, and treating its data as a sequence of uniform elements.