Editor: implemented SpriteImportColorDepth selection
This lets user to choose whether to import a sprite, converting to the game's color depth (default), or use one of the explicit options.
Currently this only lets to import indexed images keeping them as 8-bit sprite. Primary purpose is to let have 8-bit sprites used as masks in 32-bit games.